+++ src/main/java/spring1/service/HitungVolumeService.java
@@ -11,5 +11,7 @@ package spring1.service;
  */
 public interface HitungVolumeService {
 
-  double hitungLuas(double param);
+      double hitungLuas(int param);
+
+    
 }
